+++ game_states.py
@@ -3,4 +3,5 @@ from enum import Enum
 class GameStates(Enum):
     PLAYER_TURN = 1
     ENEMY_TURN = 2
+    GRAPHICAL = 3