+++ map_objects/tile.py
@@ -4,7 +4,10 @@ class Tile:
     """
     A tile on a map. It may or may not be blocked, and may or may not block sight.
     """
-    def __init__(self, blocked, palette_type, block_sight=None, palette_seed = None):
+    def __init__(self, blocked, palette_type, x, y, state = 0, block_sight=None, palette_seed = None):
+        self.x = x
+        self.y = y
+        self.state = state
         self.blocked = blocked
         self.set_palette(palette_type, palette_seed)
         # By default, if a tile is blocked, it also blocks sight
@@ -17,13 +20,16 @@ class Tile:
 
     def step(self, player):
         self.last_seen += 1
-        if self.last_seen / (4**player.stats.intelligence + self.last_seen) > random():
+        if self.last_seen / (4**player.stats.memory + self.last_seen) > random():
             self.explored = False
             self.last_seen = 0
             return(True)
         return(False)
     
-    def update_tile(self, blocked, palette_type, block_sight=None, palette_seed=None):
+    def update_tile(self, blocked, palette_type, x, y, state = 0, block_sight=None, palette_seed=None):
+        self.x = x
+        self.y = y
+        self.state = state
         self.blocked = blocked
         self.set_palette(palette_type, palette_seed)
         # By default, if a tile is blocked, it also blocks sight
@@ -70,6 +76,12 @@ class Tile:
                 "Visible" : libtcod.Color(200, 180, 50),
                 "Explored" : libtcod.Color(50, 50, 150)
             }
+            # "Burn" : {
+            #     "Unseen" : libtcod.Color(0,0,0),
+            #     "Visible" : libtcod.Color(230-self.state*13, 0+self.state*10, 0+self.state*10),
+            #     "Explored" : libtcod.Color(0,0,0)
+            # }
+
         }
         picked = palettes[palette_type]
         for key in picked.keys():