+++ render_functions.py
@@ -46,6 +46,12 @@ def draw_entity(con, entity, fov_map):
         libtcod.console_set_default_foreground(con, entity.color)
         libtcod.console_put_char(con, entity.x, entity.y, entity.char, libtcod.BKGND_NONE)
 
+def render_animations(con, animations, fov_map):
+    for ani in animations:
+        tle = ani.next_step()
+        if libtcod.map_is_in_fov(fov_map, tle.x, tle.y):
+            libtcod.console_set_char_background(con, tle.x, tle.y, tle.get_color("Visible"), libtcod.BKGND_SET)
+        
 
 def clear_entity(con, entity):
     # erase the character that represents this object